power-social-dashboard/requirements.txt
streamlit==1.36.0
plotly==5.22.0
python-dotenv==1.0.1
requests==2.32.3
google-api-python-client==2.137.0
pytz==2024.1